shorten the sub-packages in psyke/utils
