+++ psyke/regression/__init__.py
@@ -8,7 +8,6 @@ import numpy as np
 import pandas as pd
 
 from psyke import get_default_random_seed
-from psyke.regression import ZippedDimension, Limit, MinUpdate
 from psyke.regression.utils import Limit, MinUpdate, ZippedDimension, Expansion
