+++ psyke/utils/dataframe.py
@@ -1,5 +1,8 @@
+from hashlib import sha256
 from typing import Iterable
 import pandas as pd
+from pandas.core.util.hashing import hash_pandas_object
+
 from psyke import DiscreteFeature
 from psyke.schema import LessThan, GreaterThan, Between, Value
 
@@ -57,3 +60,16 @@ def get_discrete_dataset(dataset: pd.DataFrame, discrete_features: Iterable[Disc
             new_dataset[new_feature] = new_dataset[new_feature].astype(str).astype(int)
 
     return new_dataset
+
+
+class HashableDataFrame(pd.DataFrame):
+    def __init__(self, obj):
+        super().__init__(obj)
+
+    def __hash__(self):
+        hash_value = sha256(hash_pandas_object(self, index=True).values)
+        hash_value = hash(hash_value.hexdigest())
+        return hash_value
+
+    def __eq__(self, other):
+        return self.equals(other)
