--- psyke/utils/hashable.py
@@ -1,16 +1,0 @@
-from hashlib import sha256
-import pandas as pd
-from pandas.core.util.hashing import hash_pandas_object
-
-
-class HashableDataFrame(pd.DataFrame):
-    def __init__(self, obj):
-        super().__init__(obj)
-
-    def __hash__(self):
-        hash_value = sha256(hash_pandas_object(self, index=True).values)
-        hash_value = hash(hash_value.hexdigest())
-        return hash_value
-
-    def __eq__(self, other):
-        return self.equals(other)
