+++ psyke/utils/logic.py
